[BigFix] Fixes the object_count parsing in MO_BPMN_GRAPH. It can overflow int32. (#1491)
## Why I'm doing:

While testing process mining operators, MO_BPMN_GRAPH() returned an
error on walgreens_5b data model.
I found that it was because object counts in the data model were larger
than 2^31 -1 while we used int32 in parsing.

## What I'm doing:

Fixes the object_count parsing in MO_BPMN_GRAPH.

## **TODO**
In further testing, I also found that CPML
[code](https://github.com/celonis/cpm-query-engine/blob/439e57568ce2940848d4edc62cfeb64646158da7/query-engine/src/main/native/celonis-process-mining-library/include/cpml/variant/types_and_constants.h#L16-L17)
is using int32 for variant counts.
```
// A limit of int32-max should be safe to use.
using variant_count_t = std::int32_t;
```
But in the walgreens data model, we have a variant with 8.5e8 count
which is close to the 2.1e9 int32 limit.
We would need to update CPML to use int64.

## What type of PR is this:

- [X] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [X] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [X] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5

+++ be/src/exprs/celonis/modules/operators/mo/mo_bpmn_graph_helper.cpp
@@ -102,13 +102,13 @@ build(const std::string& process_tree_json, int id, const common::execution_cont
         auto process_tree_type = vertex_values[i]["process_tree_type"].GetInt();
         switch (process_tree_type) {
             case 0: {
-                size_t object_count = vertex_values[i]["object_count"].GetInt();
+                size_t object_count = vertex_values[i]["object_count"].GetInt64();
                 nodes.emplace_back(process_tree{process_tree::tau{object_count}});
                 break;
             }
             case 1: {
                 row_id activity_id = col_ptrs_ac[activity_indexes[i]];
-                size_t object_count = vertex_values[i]["object_count"].GetInt();
+                size_t object_count = vertex_values[i]["object_count"].GetInt64();
                 nodes.emplace_back(process_tree{process_tree::activity{activity_id, object_count}});
                 break;
             }
